refactor(midi): move parser and tests into @dawcore/midi
