+++ packages/dawcore-midi/src/index.ts
@@ -1,3 +1,2 @@
-// Re-exports populated in next task — keep file present so pnpm install picks up
-// the package and tsup has an entry target.
-export {};
+export { parseMidiFile, parseMidiUrl } from './parseMidiFile';
+export type { ParsedMidi, ParsedMidiTrack, ParseMidiOptions } from './parseMidiFile';
